+++ Changes.md
@@ -19,6 +19,7 @@ Change Log in version 3.0.3 (2020073104)
 16. Fix 'Install fails on Moodle 3.9' - #198 - thanks to https://gitlab.com/kiklopgs for the patch in https://gitlab.com/jezhops/moodle-theme_adaptable/-/merge_requests/34.
 17. Fix 'Gradebook: Edit link not working' - #201.
 18. Fix 'edit_button in renderers.php is not used' - #202.
+19. Fix 'Redundant CSS' - #96.
 
 Change Log in version 3.0.2 (2020073103)
 ========================================